example dialogue and enter town anim

+++ src/game/overworld/player.rs
@@ -15,7 +15,7 @@ impl Plugin for PlayerPlugin {
         app.add_event::<PlayerSpawnEvent>()
             .add_system(player_spawn)
             .add_system(player_controls.before(BoatSystems::Update))
-            .add_system(player_enter_island)
+            .add_system(player_enter_town)
             .add_system(player_camera.label(PlayerSystems::Camera))
             .add_system(player_set_attack);
     }
@@ -25,7 +25,9 @@ impl Plugin for PlayerPlugin {
 pub struct PlayerSpawnEvent;
 
 #[derive(Component)]
-pub struct Player;
+pub struct Player {
+    disabled: bool,
+}
 
 fn player_spawn(
     mut ev_spawn: EventReader<PlayerSpawnEvent>,
@@ -36,7 +38,7 @@ fn player_spawn(
     for _ in ev_spawn.iter() {
         let entity = commands
             .spawn()
-            .insert(Player)
+            .insert(Player { disabled: false })
             .insert(Label("Player".to_owned()))
             .insert(AudioPlusListener)
             .id();
@@ -50,15 +52,20 @@ fn player_spawn(
 }
 
 fn player_controls(
-    mut query: Query<(&mut Boat, &GlobalTransform), With<Player>>,
+    mut query: Query<(&mut Boat, &GlobalTransform, &Player)>,
     mouse: Res<Mouse>,
     input: Res<Input<MouseButton>>,
     keys: Res<Input<KeyCode>>,
+    cutscenes: Res<Cutscenes>,
 ) {
     if query.is_empty() {
         return;
     }
-    for (mut boat, global_transform) in query.iter_mut() {
+    for (mut boat, global_transform, player) in query.iter_mut() {
+        if player.disabled || cutscenes.running() {
+            boat.movement = Vec2::ZERO;
+            continue;
+        }
         let mut mouse_aim = (mouse.position - global_transform.translation().truncate()) / 200.;
         if mouse_aim.length_squared() == 0. {
             mouse_aim = Vec2::new(0.1, 0.);
@@ -77,28 +84,36 @@ fn player_controls(
     }
 }
 
-fn player_enter_island(
-    mut app_state: ResMut<State<AppState>>,
+fn player_enter_town(
     mut game_state: ResMut<GameState>,
     island_query: Query<(Entity, &Town)>,
-    player_query: Query<Entity, With<Player>>,
+    mut player_query: Query<(Entity, &mut Player)>,
     transform_query: Query<&GlobalTransform>,
+    mut ev_cutscene_enter_town: EventWriter<CutsceneStartEvent<EnterTownCutscene>>,
 ) {
-    'outer: for (island_entity, island) in island_query.iter() {
-        let island_position = if let Ok(island_transform) = transform_query.get(island_entity) {
-            island_transform.translation().truncate()
+    'outer: for (town_entity, island) in island_query.iter() {
+        let town_position = if let Ok(town_transform) = transform_query.get(town_entity) {
+            town_transform.translation().truncate()
         } else {
             continue;
         };
-        for player_entity in player_query.iter() {
+        for (player_entity, mut player) in player_query.iter_mut() {
+            if player.disabled {
+                continue;
+            }
             let player_position = if let Ok(player_transform) = transform_query.get(player_entity) {
                 player_transform.translation().truncate()
             } else {
                 continue;
             };
-            if player_position.distance(island_position) < 200. {
+            if player_position.distance(town_position) < 200. {
+                player.disabled = true;
+                ev_cutscene_enter_town.send(CutsceneStartEvent(EnterTownCutscene {
+                    boat: Some(player_entity),
+                    from: player_position,
+                    to: town_position + Vec2::new(-10., -100.),
+                }));
                 game_state.town = island.town.clone();
-                app_state.set(AppState::TownOutside).unwrap();
                 break 'outer;
             }
         }
